fosse/chemical-equipment-app-new/requirements.txt
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.2.0
djangorestframework-simplejwt==5.3.1
reportlab==4.0.9
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
